GenomicElementType

Documentation for GenomicElementType class from SLiM

GenomicElementType

Format

An object of class list of length 10.

Details

This class has the following methods (functions):

getValue

setMutationFractions

setMutationMatrix

setValue

This class has the following properties:

color

A property of type string. It is of length one (a singleton). This property is a variable, so it is modifiable. Property Description: The color used to display genomic elements of this type in SLiMgui. Outside of SLiMgui, this property still exists, but is not used by SLiM. Colors may be specified by name, or with hexadecimal RGB values of the form "#RRGGBB" (see the Eidos manual). If color is the empty string, "", SLiMgui’s default color scheme is used; this is the default for new GenomicElementType objects.

id

A property of type integer. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The identifier for this genomic element type; for genomic element type g3, for example, this is 3.

mutationFractions

A property of type float. This property is a constant, so it is not modifiable. Property Description: For each MutationType represented in this genomic element type, this property has the corresponding fraction of all mutations that will be drawn from that MutationType.

mutationMatrix

A property of type float. This property is a constant, so it is not modifiable. Property Description: The nucleotide mutation matrix used for this genomic element type, set up by initializeGenomicElementType() and setMutationMatrix(). This property is only defined in nucleotide-based models; it is unavailable otherwise.

mutationTypes

A property of type MutationType object. This property is a constant, so it is not modifiable. Property Description: The MutationType instances used by this genomic element type.

tag

A property of type integer. It is of length one (a singleton). This property is a variable, so it is modifiable. Property Description: A user-defined integer value. The value of tag is initially undefined, and it is an error to try to read it; if you wish it to have a defined value, you must arrange that yourself by explicitly setting its value prior to using it elsewhere in your code. The value of tag is not used by SLiM; it is free for you to use. See also the getValue() and setValue() methods, for another way of attaching state to genomic element types.