SLiMSim

Documentation for SLiMSim class from SLiM

SLiMSim

Format

An object of class list of length 48.

Details

This class has the following methods (functions):

addSubpop

addSubpopSplit

countOfMutationsOfType

deregisterScriptBlock

getValue

mutationCounts

mutationFrequencies

mutationsOfType

outputFixedMutations

outputFull

outputMutations

outputUsage

readFromPopulationFile

recalculateFitness

registerEarlyEvent

registerFitnessCallback

registerInteractionCallback

registerLateEvent

registerMateChoiceCallback

registerModifyChildCallback

registerRecombinationCallback

registerMutationCallback

registerReproductionCallback

rescheduleScriptBlock

setValue

simulationFinished

treeSeqCoalesced

treeSeqOutput

treeSeqRememberIndividuals

treeSeqSimplify

This class has the following properties:

chromosome

A property of type Chromosome object. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The Chromosome object used by the simulation.

chromosomeType

A property of type string. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The type of chromosome being simulated; this will be one of "A", "X", or "Y".

dimensionality

A property of type string. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The spatial dimensionality of the simulation, as specified in initializeSLiMOptions(). This will be "" (the empty string) for non-spatial simulations (the default), or "x", "xy", or "xyz", for simulations using those spatial dimensions respectively.

dominanceCoeffX

A property of type float. It is of length one (a singleton). This property is a variable, so it is modifiable. Property Description: The dominance coefficient value used to modify the selection coefficients of mutations present on the single X chromosome of an XY male (see the SLiM documentation for details). Used only when simulating an X chromosome; setting a value for this property in other circumstances is an error. Changing this will normally affect the fitness values calculated at the end of the current generation; if you want current fitness values to be affected, you can call SLiMSim’s method recalculateFitness() – but see the documentation of that method for caveats.

generation

A property of type integer. It is of length one (a singleton). This property is a variable, so it is modifiable. Property Description: The current generation number.

genomicElementTypes

A property of type GenomicElementType object. This property is a constant, so it is not modifiable. Property Description: The GenomicElementType objects being used in the simulation.

inSLiMgui

A property of type logical. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: This property has been deprecated, and may be removed in a future release of SLiM. In SLiM 3.2.1 and later, use exists("slimgui") instead. If T, the simulation is presently running inside SLiMgui; if F, it is running at the command line. In general simulations should not care where they are running, but in special circumstances such as opening plot windows it may be necessary to know the runtime environment.

interactionTypes

A property of type InteractionType object. This property is a constant, so it is not modifiable. Property Description: The InteractionType objects being used in the simulation.

modelType

A property of type string. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The type of model being simulated, as specified in initializeSLiMModelType(). This will be "WF" for WF models (Wright–Fisher models, the default), or "nonWF" for nonWF models (non-Wright–Fisher models; see section 1.6 for discussion).

mutationTypes

A property of type MutationType object. This property is a constant, so it is not modifiable. Property Description: The MutationType objects being used in the simulation.

mutations

A property of type Mutation object. This property is a constant, so it is not modifiable. Property Description: The Mutation objects that are currently active in the simulation.

nucleotideBased

A property of type logical. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: If T, the model is nucleotide-based; if F, it is not. See the discussion of the nucleotideBased parameter to initializeSLiMOptions() for discussion.

periodicity

A property of type string. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The spatial periodicity of the simulation, as specified in initializeSLiMOptions(). This will be "" (the empty string) for non- spatial simulations and simulations with no periodic spatial dimensions (the default). Otherwise, it will be a string representing the subset of spatial dimensions that have been declared to be periodic, as specified to initializeSLiMOptions().

scriptBlocks

A property of type SLiMEidosBlock object. This property is a constant, so it is not modifiable. Property Description: All registered SLiMEidosBlock objects in the simulation.

sexEnabled

A property of type logical. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: If T, sex is enabled in the simulation; if F, individuals are hermaphroditic.

subpopulations

A property of type Subpopulation object. This property is a constant, so it is not modifiable. Property Description: The Subpopulation instances currently defined in the simulation.

substitutions

A property of type Substitution object. This property is a constant, so it is not modifiable. Property Description: A vector of Substitution objects, representing all mutations that have been fixed.

tag

A property of type integer. It is of length one (a singleton). This property is a variable, so it is modifiable. Property Description: A user-defined integer value. The value of tag is initially undefined, and it is an error to try to read it; if you wish it to have a defined value, you must arrange that yourself by explicitly setting its value prior to using it elsewhere in your code. The value of tag is not used by SLiM; it is free for you to use. See also the getValue() and setValue() methods, for another way of attaching state to the simulation.