slimrlang stub for the first part of the SLiM ternary operator (condition ? yes else no).

This is used in conjunction with which will make the code valid in R. Note that the R function is a stub, it does not do anything in R (except bring up this documentation). It will only do anything useful when used inside a \code\linkslim_block function further nested in a \code\linkslim_script function call, where it will be translated into valid SLiM code as part of a full SLiM script.

lhs %?% rhs

Arguments

lhs

A condition

rhs

A SLiM expression to be executed if the condition is TRUE

Copyright

This is documentation for a function in the SLiM software, and has been reproduced from the official manual, which can be found here: \urlhttp://benhaller.com/slim/SLiM_Manual.pdf. This documentation is Copyright © 2016–2020 Philipp Messer. All rights reserved. More information about SLiM can be found on the official website: \urlhttps://messerlab.org/slim/

Author

Benjamin C Haller (\emailbhaller@benhaller.com) and Philipp W Messer (\emailmesser@cornell.edu)