SLiM method initializeSex

Documentation for SLiM function initializeSex, which is a method of the SLiM class Initialize. Note that the R function is a stub, it does not do anything in R (except bring up this documentation). It will only do anything useful when used inside a slim_block function further nested in a slim_script function call, where it will be translated into valid SLiM code as part of a full SLiM script.

initializeSex(chromosomeType, xDominanceCoeff)

Arguments

chromosomeType
An object of type string or numeric. Must be of length 1 (a singleton). See details for description.

xDominanceCoeff
An object of type string or numeric. Must be of length 1 (a singleton). The default value is 1. See details for description.

Value

An object of type void.

Details

Documentation for this function can be found in the official SLiM manual: page 523.

Enable and configure sex in the simulation. The argument chromosomeType gives the type of chromosome to be simulated; this should be "A", "X", or "Y". If the chromosomeType is "X", the optional xDominanceCoeff parameter can supply the dominance coefficient used when a mutation is present in an XY male, and is thus "heterozygous" (but in a different sense than the heterozygosity of an XX female with one copy of the mutation). Calling this function has the side effect of enabling sex in the simulation; individuals will be male and female (rather than hermaphroditic) regardless of the chromosomeType chosen for simulation. There is no way to disable sex once it has been enabled; if you don't want to have sex, don't call this function.

Copyright

This is documentation for a function in the SLiM software, and has been reproduced from the official manual, which can be found here: http://benhaller.com/slim/SLiM_Manual.pdf. This documentation is Copyright © 2016-2020 Philipp Messer. All rights reserved. More information about SLiM can be found on the official website: https://messerlab.org/slim/

See also

Author

Benjamin C Haller (bhaller@benhaller.com) and Philipp W Messer (messer@cornell.edu)