SLiM method interactionDistance

Documentation for SLiM function interactionDistance, which is a method of the SLiM class InteractionType. Note that the R function is a stub, it does not do anything in R (except bring up this documentation). It will only do anything useful when used inside a slim_block function further nested in a slim_script function call, where it will be translated into valid SLiM code as part of a full SLiM script.

interactionDistance(receiver, exerters)

Arguments

receiver
An object of type Individual object. Must be of length 1 (a singleton). See details for description.

exerters
An object of type null or Individual object. The default value is NULL. See details for description.

Value

An object of type float.

Details

Returns a vector containing interaction-dependent distances between receiver and individuals in exerters that exert an interaction strength upon receiver. If exerters is NULL (the default), then a vector of the interaction- dependent distances from receiver to all individuals in its subpopulation (including receiver itself) is returned; this case may be handled much more efficiently than if a vector of all individuals in the subpopulation is explicitly provided. If the InteractionType is nonspatial, this method may not be called. Importantly, distances are calculated according to the spatiality of the InteractionType (as declared in initializeInteractionType()), not the dimensionality of the model as a whole (as declared in initializeSLiMOptions()). The distances returned are therefore the distances that would be used to calculate interaction strengths. In addition, interactionDistance() will return INF as the distance between receiver and any individual which does not exert an interaction upon receiver; the interactionDistance() between an individual and itself will thus be INF, and likewise for pairs excluded from interacting by the sex segregation or max distance of the interaction type. See distance() for an alternative distance definition.

Copyright

This is documentation for a function in the SLiM software, and has been reproduced from the official manual, which can be found here: http://benhaller.com/slim/SLiM_Manual.pdf. This documentation is Copyright © 2016–2020 Philipp Messer. All rights reserved. More information about SLiM can be found on the official website: https://messerlab.org/slim/

Author

Benjamin C Haller (bhaller@benhaller.com) and Philipp W Messer (messer@cornell.edu)