SLiM method mutationFrequencies

Documentation for SLiM function mutationFrequencies, which is a method of the SLiM class SLiMSim. Note that the R function is a stub, it does not do anything in R (except bring up this documentation). It will only do anything useful when used inside a slim_block function further nested in a slim_script function call, where it will be translated into valid SLiM code as part of a full SLiM script.

mutationFrequencies(subpops, mutations)

Arguments

subpops
An object of type null or Subpopulation object. See details for description.

mutations
An object of type null or Mutation object. The default value is NULL. See details for description.

Value

An object of type float.

Details

Documentation for this function can be found in the official SLiM manual: page 567.

Return a float vector with the frequencies of all of the Mutation objects passed in mutations, within the Subpopulation objects in subpops. The subpops argument is required, but you may pass NULL to get population- wide frequencies. If the optional mutations argument is NULL (the default), frequencies will be returned for all of the active Mutation objects in the simulation - the same Mutation objects, and in the same order, as would be returned by the mutations property of sim, in other words. See the -mutationCounts() method to obtain integer counts instead of float frequencies. See also the Genome methods mutationCountsInGenomes() and mutationFrequenciesInGenomes().

Copyright

This is documentation for a function in the SLiM software, and has been reproduced from the official manual, which can be found here: http://benhaller.com/slim/SLiM_Manual.pdf. This documentation is Copyright © 2016-2020 Philipp Messer. All rights reserved. More information about SLiM can be found on the official website: https://messerlab.org/slim/

See also

Author

Benjamin C Haller (bhaller@benhaller.com) and Philipp W Messer (messer@cornell.edu)