SLiM method registerRecombinationCallback

Documentation for SLiM function registerRecombinationCallback, which is a method of the SLiM class SLiMSim. Note that the R function is a stub, it does not do anything in R (except bring up this documentation). It will only do anything useful when used inside a slim_block function further nested in a slim_script function call, where it will be translated into valid SLiM code as part of a full SLiM script.

registerRecombinationCallback(id, source, subpop, start, end)

Arguments

id
An object of type null or integer or string. Must be of length 1 (a singleton). See details for description.

source
An object of type string. Must be of length 1 (a singleton). See details for description.

subpop
An object of type null or integer or Subpopulation object. Must be of length 1 (a singleton). The default value is NULL. See details for description.

start
An object of type null or integer. Must be of length 1 (a singleton). The default value is NULL. See details for description.

end
An object of type null or integer. Must be of length 1 (a singleton). The default value is NULL. See details for description.

Value

An object of type SLiMEidosBlock object. Return will be of length 1 (a singleton)

Details

Documentation for this function can be found in the official SLiM manual: page 573.

Register a block of Eidos source code, represented as the string singleton source, as an Eidos recombination() callback in the current simulation, with optional subpopulation subpop (which may be an integer identifier, or NULL, the default, to indicate all subpopulations) and optional start and end generations all limiting its applicability. The script block will be given identifier id (specified as an integer, or as a string symbolic name such as "s5"); this may be NULL if there is no need to be able to refer to the block later. The registered callback is added to the end of the list of registered SLiMEidosBlock objects, and is active immediately; it may be eligible to execute in the current generation (see section 25.9 for details). The new SLiMEidosBlock will be defined as a global variable immediately by this method (see section 24.11), and will also be returned by this method.

Copyright

This is documentation for a function in the SLiM software, and has been reproduced from the official manual, which can be found here: http://benhaller.com/slim/SLiM_Manual.pdf. This documentation is Copyright © 2016-2020 Philipp Messer. All rights reserved. More information about SLiM can be found on the official website: https://messerlab.org/slim/

See also

Author

Benjamin C Haller (bhaller@benhaller.com) and Philipp W Messer (messer@cornell.edu)