Run a SLiM script from R

This function runs a SLiM script, specified as a slimr_script object, a character vector, or a text file.

slim_run(
  x,
  slim_path = NULL,
  script_file = NULL,
  simple_run = FALSE,
  capture_output = TRUE,
  show_output = !capture_output,
  callbacks = NULL,
  new_grdev = FALSE,
  ...
)

# S3 method for character
slim_run(x, ...)

# S3 method for slimr_script
slim_run(x, ...)

Arguments

callbacks
A list of functions to be called during the SLiM run. This can be used to dynamically transform or visualise output from the simulation while it is running. It should be of the form function(data, ...) {do something..}. If using slimr_output to get formatted data output from SLiM, data will be a four column tibble containing output from the current iteration of the simulation. Columns are:

generation

A vector of generations processed in the current iteration

name

Names of the output data.

expression

The SLiM expression used to generate the output

data

The raw data output from SLiM as a character vector

new_grdev
Should a new graphics device window be opened on RStudio? This is mainly useful if you are using custom callbacks that generate live figures, and want a faster plotting experience. This is because the default plot viewer in RStudio can be quite slow.

...
Additional arguments to be passed to any callback functions.

Value

A slimr_results object which has the following components:

output

A character vector of raw output. Will be NULL if capture_output is FALSE

exit_status

The exit status code returned by the SLiM process. 0 means success.

Methods (by class)

character: Run a SLiM script from character vector

slimr_script: Run a SLiM script from slimr_script object