Run a SLiM script and optionally return output from it to R.

Run a SLiM script and optionally return output from it to R.

slim_run_script(
  slim_script = NULL,
  script_file = NULL,
  slim_path = NULL,
  output = NULL,
  output_every = 1,
  output_sample = NULL,
  output_format = c("genlight", "slim_tibble", "vcf_tibble"),
  .progress = TRUE,
  write_data_live = FALSE,
  log_output_to = NULL
)

Arguments

slim_script
This can either be a character vector (of length 1) containing the SLiM script to run or a slim_script object such as created by slim_script_from_text. If this is NULL then you must provide a text file with your script in script_file.

script_file
A file path specifying a text file with a valid SLiM script to run. If you specify both this and slim_script, slim_script will be run and this parameter will be ignored.

slim_path
Path to SLiM executable. If NULL slimr will try and find the SLiM executable through default locations or environmental variables. See documentation for slim_setup() for details.

output
What would you like to output from SLiM to R? A character vector containing the names of output to generate. See details for valid output names.

output_every
How often should R collect output from SLiM? Output will be collected once every output_every generations.

output_sample
When generating output from SLiM, how many samples of the population to take? 0 means sample all extant members of the population.

output_format
In what format should the genomic data be outputted? Valid values are: "VCF", "genlight", or "tibble". Regardless of your choice the genomes will be contained in a list column of the slim_sim tibble output (named "genomes".

.progress
Should a progress bar be displayed while the SLiM script is running?

write_data_live
Should the output being read into R from SLiM be saved as it is running? Use log_output_to to specify a file to save to.

log_output_to
File path to save output to. Ignored if write_data_live is FALSE

Value

A slim_sim object with output from the simulation if output is specified. If output is not specified or if slim_script is of class character, this will be the exit code of the SLiM run (exit code 0 means it ran successfully)

Details

Valid output names that can be added to the output parameter include:

"genomes""Add this to get output on simulated genomes."

"individuals"Add this for output on individuals, specifically their (two) genomes."

"coordinates"If running a spatially explicit simulation, add this to output spatial coordinates of individuals."