Render a SLiM script with special slimrlang formatting

If your slimr_script object has made use of special slimrlang syntax slimr_template, this function will 'render' the slimr_script into valid SLiM syntax, ready to be run with SLiM or slim_run

slim_script_render(
  slimr_script,
  template = NULL,
  replace_NAs = TRUE,
  reps = 1,
  parallel = FALSE
)

Arguments

slimr_script

The slimr_script object to be rendered

template

A list or data.frame containing values for any templated variables. If a list, it must be named, where the names correspond to the variables. If a list of lists, the inner lists must be named with the variable names, and slim_script_render will render a separate slimr_script for each top-level list element and return it as a slimr_script_coll object. If a data.frame (or tibble), then the column names should match the templated variables, and slim_script_render will render a separate slimr_script for each row and return it as a slimr_script_coll object.

replace_NAs

Should NA values in the template be replaced by their default values?

reps

Should the rendered script be replicated? If greater than 1, a slimr_script_coll will be returned. This can also be used with a slimr_script object that has already been rendered, in which case it will just repeat the rendered script in the result.

parallel

Should the rendering be done in parallel when rendering multiple scripts? Requires the furrr package and will use the plan set by future::plan