SLiMEidosBlock

Documentation for SLiMEidosBlock class from SLiM

SLiMEidosBlock

Format

An object of class list of length 7.

Details

This class represents a block of Eidos code registered in a SLiM simulation. All Eidos events and Eidos callbacks defined in the SLiM input file of the current simulation are instantiated as SLiMEidosBlock objects and are available through the scriptBlocks property of SLiMSim; see section 24.13.1. In addition, new script blocks can be created programmatically and registered with the simulation, and registered script blocks can be deregistered; see the ‑register...() and ‑deregisterScriptBlock() methods of SLiMSim in section 24.13.2. The currently executing script block is available through the self global; see section 25.9. This class has the following methods (functions):

None. This class has no methods.

This class has the following properties:

active

A property of type integer or integer or integer or string or integer or integer or string. It is of length one (a singleton). This property is a variable, so it is modifiable. Property Description: If this evaluates to logical F (i.e., is equal to 0), the script block is inactive and will not be called. The value of active for all registered script blocks is reset to -1 at the beginning of each generation, prior to script events being called, thus activating all blocks. Any integer value other than -1 may be used instead of -1 to represent that a block is active; for example, active may be used as a counter to make a block execute a fixed number of times in each generation. This value is not cached by SLiM; if it is changed, the new value takes effect immediately. For example, a callback might be activated and inactivated repeatedly during a single generation.

end

A property of type integer or integer or integer or string or integer or integer or string. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The last generation in which the script block is active.

id

A property of type integer or integer or integer or string or integer or integer or string. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The identifier for this script block; for script s3, for example, this is 3. A script block for which no id was given will have an id of -1.

source

A property of type integer or integer or integer or string or integer or integer or string. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The source code string of the script block.

start

A property of type integer or integer or integer or string or integer or integer or string. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The first generation in which the script block is active.

tag

A property of type integer or integer or integer or string or integer or integer or string. It is of length one (a singleton). This property is a variable, so it is modifiable. Property Description: A user-defined integer value. The value of tag is initially undefined, and it is an error to try to read it; if you wish it to have a defined value, you must arrange that yourself by explicitly setting its value prior to using it elsewhere in your code. The value of tag is not used by SLiM; it is free for you to use.

type

A property of type integer or integer or integer or string or integer or integer or string. It is of length one (a singleton). This property is a constant, so it is not modifiable. Property Description: The type of the script block; this will be "early" or "late" for the two types of Eidos events, or "initialize", "fitness", "mateChoice", "modifyChild", or "recombination" for the respective types of Eidos callbacks (see section 24.1 and chapter 25).