Create a file to initialise a population in SLiM

Make a SLiM file in the style of outputFull() which can be read by SLiM. This is useful to start a simulation with a particular population state that you define in R

slim_make_pop_input(
  snps,
  file_name = tempfile(),
  sim_gen = 10000,
  ind_pops = NULL,
  ind_sex = NULL,
  mut_pos = NULL,
  mut_type = NULL,
  mut_sel = NULL,
  mut_dom = NULL,
  mut_pop = NULL,
  mut_gen = NULL,
  mut_nuc = NULL,
  version = 4
)

Arguments

snps
R object containing SNPs. Can be a matrix with 0, 1, or 2 as its elements, or a genlight object. Rows correspond to individuals, columns to loci.

file_name
Path to file where the population input file should be saved.

sim_gen
What generation to write in the file. This doesn't really do anything, but is just a requirement for the format.

ind_pops
An optional character vector with length equal to nrow(snps) with Subpopulation indicators for each individual. Should be "p1", "p2", etc. as this is how subpopulations are named in SLiM.

ind_sex
An optional character vector with length equal to nrow(snps) with the sex for each individual (can be "H" for hermaphrodite, "F" for female, or "M" for male.)

mut_pos
An optional integer vector with length equal to ncol(snps) specifying the positions in the genome of each loci.

mut_type
An optional character vector with length equal to nrow(snps) with the mutation type of each loci. Should be "m1", "m2", etc. as this is how mutation types are named in SLiM.

mut_sel
An optional numeric vector with length equal to nrow(snps) with the selection coefficient for the mutation type of each loci.

mut_dom
An optional numeric vector with length equal to nrow(snps) with the dominance coefficient for the mutation type of each loci.

mut_pop
An optional character vector with length equal to nrow(snps) with the subpopulation of origin for the mutation at each loci. Should be "p1", "p2", etc. as this is how subpopulations are named in SLiM.

mut_gen
An optional integer vector with length equal to ncol(snps) specifying the generation of origin for the mutation at each loci.

mut_nuc
An optional character vector with length equal to nrow(snps) with the nucleotide of the mutation. Should be "A", "C", "G", or "T".

version
SLiM output version number to use.

Value

Returns the file name where the population data was saved

Examples

pop_file <- slim_make_pop_input(matrix(rbinom(1000, 2, 0.25), nrow = 100, ncol = 100))
cat(readLines(pop_file))

#> #OUT 10000 A Version: 4 Populations: p1 100 H Mutations: 0 0 m1 1 0 0.5 p1 1 50 1 1 m1 2 0 0.5 p1 1 63 2 2 m1 3 0 0.5 p1 1 57 3 3 m1 4 0 0.5 p1 1 57 4 4 m1 5 0 0.5 p1 1 45 5 5 m1 6 0 0.5 p1 1 58 6 6 m1 7 0 0.5 p1 1 53 7 7 m1 8 0 0.5 p1 1 47 8 8 m1 9 0 0.5 p1 1 47 9 9 m1 10 0 0.5 p1 1 45 10 10 m1 11 0 0.5 p1 1 50 11 11 m1 12 0 0.5 p1 1 63 12 12 m1 13 0 0.5 p1 1 57 13 13 m1 14 0 0.5 p1 1 57 14 14 m1 15 0 0.5 p1 1 45 15 15 m1 16 0 0.5 p1 1 58 16 16 m1 17 0 0.5 p1 1 53 17 17 m1 18 0 0.5 p1 1 47 18 18 m1 19 0 0.5 p1 1 47 19 19 m1 20 0 0.5 p1 1 45 20 20 m1 21 0 0.5 p1 1 50 21 21 m1 22 0 0.5 p1 1 63 22 22 m1 23 0 0.5 p1 1 57 23 23 m1 24 0 0.5 p1 1 57 24 24 m1 25 0 0.5 p1 1 45 25 25 m1 26 0 0.5 p1 1 58 26 26 m1 27 0 0.5 p1 1 53 27 27 m1 28 0 0.5 p1 1 47 28 28 m1 29 0 0.5 p1 1 47 29 29 m1 30 0 0.5 p1 1 45 30 30 m1 31 0 0.5 p1 1 50 31 31 m1 32 0 0.5 p1 1 63 32 32 m1 33 0 0.5 p1 1 57 33 33 m1 34 0 0.5 p1 1 57 34 34 m1 35 0 0.5 p1 1 45 35 35 m1 36 0 0.5 p1 1 58 36 36 m1 37 0 0.5 p1 1 53 37 37 m1 38 0 0.5 p1 1 47 38 38 m1 39 0 0.5 p1 1 47 39 39 m1 40 0 0.5 p1 1 45 40 40 m1 41 0 0.5 p1 1 50 41 41 m1 42 0 0.5 p1 1 63 42 42 m1 43 0 0.5 p1 1 57 43 43 m1 44 0 0.5 p1 1 57 44 44 m1 45 0 0.5 p1 1 45 45 45 m1 46 0 0.5 p1 1 58 46 46 m1 47 0 0.5 p1 1 53 47 47 m1 48 0 0.5 p1 1 47 48 48 m1 49 0 0.5 p1 1 47 49 49 m1 50 0 0.5 p1 1 45 50 50 m1 51 0 0.5 p1 1 50 51 51 m1 52 0 0.5 p1 1 63 52 52 m1 53 0 0.5 p1 1 57 53 53 m1 54 0 0.5 p1 1 57 54 54 m1 55 0 0.5 p1 1 45 55 55 m1 56 0 0.5 p1 1 58 56 56 m1 57 0 0.5 p1 1 53 57 57 m1 58 0 0.5 p1 1 47 58 58 m1 59 0 0.5 p1 1 47 59 59 m1 60 0 0.5 p1 1 45 60 60 m1 61 0 0.5 p1 1 50 61 61 m1 62 0 0.5 p1 1 63 62 62 m1 63 0 0.5 p1 1 57 63 63 m1 64 0 0.5 p1 1 57 64 64 m1 65 0 0.5 p1 1 45 65 65 m1 66 0 0.5 p1 1 58 66 66 m1 67 0 0.5 p1 1 53 67 67 m1 68 0 0.5 p1 1 47 68 68 m1 69 0 0.5 p1 1 47 69 69 m1 70 0 0.5 p1 1 45 70 70 m1 71 0 0.5 p1 1 50 71 71 m1 72 0 0.5 p1 1 63 72 72 m1 73 0 0.5 p1 1 57 73 73 m1 74 0 0.5 p1 1 57 74 74 m1 75 0 0.5 p1 1 45 75 75 m1 76 0 0.5 p1 1 58 76 76 m1 77 0 0.5 p1 1 53 77 77 m1 78 0 0.5 p1 1 47 78 78 m1 79 0 0.5 p1 1 47 79 79 m1 80 0 0.5 p1 1 45 80 80 m1 81 0 0.5 p1 1 50 81 81 m1 82 0 0.5 p1 1 63 82 82 m1 83 0 0.5 p1 1 57 83 83 m1 84 0 0.5 p1 1 57 84 84 m1 85 0 0.5 p1 1 45 85 85 m1 86 0 0.5 p1 1 58 86 86 m1 87 0 0.5 p1 1 53 87 87 m1 88 0 0.5 p1 1 47 88 88 m1 89 0 0.5 p1 1 47 89 89 m1 90 0 0.5 p1 1 45 90 90 m1 91 0 0.5 p1 1 50 91 91 m1 92 0 0.5 p1 1 63 92 92 m1 93 0 0.5 p1 1 57 93 93 m1 94 0 0.5 p1 1 57 94 94 m1 95 0 0.5 p1 1 45 95 95 m1 96 0 0.5 p1 1 58 96 96 m1 97 0 0.5 p1 1 53 97 97 m1 98 0 0.5 p1 1 47 98 98 m1 99 0 0.5 p1 1 47 99 99 m1 100 0 0.5 p1 1 45 Individuals: p1:i0 H p1:0 p1:1 p1:i1 H p1:2 p1:3 p1:i2 H p1:4 p1:5 p1:i3 H p1:6 p1:7 p1:i4 H p1:8 p1:9 p1:i5 H p1:10 p1:11 p1:i6 H p1:12 p1:13 p1:i7 H p1:14 p1:15 p1:i8 H p1:16 p1:17 p1:i9 H p1:18 p1:19 p1:i10 H p1:20 p1:21 p1:i11 H p1:22 p1:23 p1:i12 H p1:24 p1:25 p1:i13 H p1:26 p1:27 p1:i14 H p1:28 p1:29 p1:i15 H p1:30 p1:31 p1:i16 H p1:32 p1:33 p1:i17 H p1:34 p1:35 p1:i18 H p1:36 p1:37 p1:i19 H p1:38 p1:39 p1:i20 H p1:40 p1:41 p1:i21 H p1:42 p1:43 p1:i22 H p1:44 p1:45 p1:i23 H p1:46 p1:47 p1:i24 H p1:48 p1:49 p1:i25 H p1:50 p1:51 p1:i26 H p1:52 p1:53 p1:i27 H p1:54 p1:55 p1:i28 H p1:56 p1:57 p1:i29 H p1:58 p1:59 p1:i30 H p1:60 p1:61 p1:i31 H p1:62 p1:63 p1:i32 H p1:64 p1:65 p1:i33 H p1:66 p1:67 p1:i34 H p1:68 p1:69 p1:i35 H p1:70 p1:71 p1:i36 H p1:72 p1:73 p1:i37 H p1:74 p1:75 p1:i38 H p1:76 p1:77 p1:i39 H p1:78 p1:79 p1:i40 H p1:80 p1:81 p1:i41 H p1:82 p1:83 p1:i42 H p1:84 p1:85 p1:i43 H p1:86 p1:87 p1:i44 H p1:88 p1:89 p1:i45 H p1:90 p1:91 p1:i46 H p1:92 p1:93 p1:i47 H p1:94 p1:95 p1:i48 H p1:96 p1:97 p1:i49 H p1:98 p1:99 p1:i50 H p1:100 p1:101 p1:i51 H p1:102 p1:103 p1:i52 H p1:104 p1:105 p1:i53 H p1:106 p1:107 p1:i54 H p1:108 p1:109 p1:i55 H p1:110 p1:111 p1:i56 H p1:112 p1:113 p1:i57 H p1:114 p1:115 p1:i58 H p1:116 p1:117 p1:i59 H p1:118 p1:119 p1:i60 H p1:120 p1:121 p1:i61 H p1:122 p1:123 p1:i62 H p1:124 p1:125 p1:i63 H p1:126 p1:127 p1:i64 H p1:128 p1:129 p1:i65 H p1:130 p1:131 p1:i66 H p1:132 p1:133 p1:i67 H p1:134 p1:135 p1:i68 H p1:136 p1:137 p1:i69 H p1:138 p1:139 p1:i70 H p1:140 p1:141 p1:i71 H p1:142 p1:143 p1:i72 H p1:144 p1:145 p1:i73 H p1:146 p1:147 p1:i74 H p1:148 p1:149 p1:i75 H p1:150 p1:151 p1:i76 H p1:152 p1:153 p1:i77 H p1:154 p1:155 p1:i78 H p1:156 p1:157 p1:i79 H p1:158 p1:159 p1:i80 H p1:160 p1:161 p1:i81 H p1:162 p1:163 p1:i82 H p1:164 p1:165 p1:i83 H p1:166 p1:167 p1:i84 H p1:168 p1:169 p1:i85 H p1:170 p1:171 p1:i86 H p1:172 p1:173 p1:i87 H p1:174 p1:175 p1:i88 H p1:176 p1:177 p1:i89 H p1:178 p1:179 p1:i90 H p1:180 p1:181 p1:i91 H p1:182 p1:183 p1:i92 H p1:184 p1:185 p1:i93 H p1:186 p1:187 p1:i94 H p1:188 p1:189 p1:i95 H p1:190 p1:191 p1:i96 H p1:192 p1:193 p1:i97 H p1:194 p1:195 p1:i98 H p1:196 p1:197 p1:i99 H p1:198 p1:199 Genomes: p1:0 A 0 1 2 4 7 10 12 14 20 24 27 30 31 32 34 37 40 41 42 44 47 50 52 54 57 60 61 64 70 71 74 77 80 82 84 87 90 94 p1:1 A 0 4 10 11 14 17 20 21 22 24 30 34 40 44 50 51 54 60 62 64 67 70 72 74 80 81 84 90 91 92 94 97 p1:2 A 1 11 19 21 29 39 41 49 51 69 71 79 89 91 99 p1:3 A 9 31 59 61 81 p1:4 A 0 1 3 13 15 19 20 21 23 25 26 30 35 36 39 46 53 59 60 63 70 83 89 90 91 p1:5 A 5 6 9 10 11 16 29 31 33 40 41 43 45 49 50 51 55 56 61 65 66 69 71 73 75 76 79 80 81 85 86 93 95 96 99 p1:6 A 10 13 20 23 37 43 47 57 60 67 70 73 77 90 97 p1:7 A 0 3 7 17 27 30 33 40 50 53 63 80 83 87 93 p1:8 A 3 13 16 23 32 33 36 43 46 52 53 56 63 66 71 73 76 81 83 91 92 93 p1:9 A 1 2 3 6 11 12 13 21 22 23 26 31 33 41 42 43 51 53 61 62 63 72 73 82 83 86 93 96 p1:10 A 0 4 9 10 11 14 19 20 21 22 24 26 27 29 30 31 34 37 39 40 42 44 46 47 49 50 51 52 56 59 60 61 62 64 67 69 70 79 80 82 84 86 87 89 90 91 92 94 96 97 99 p1:11 A 0 1 2 6 7 9 10 12 16 17 19 20 29 30 32 36 39 40 41 49 50 54 57 59 60 66 69 70 71 72 74 76 77 79 80 81 89 90 99 p1:12 A 2 16 20 28 29 30 39 48 49 56 58 62 70 79 80 88 92 96 99 p1:13 A 0 6 8 9 10 12 18 19 22 26 32 36 38 40 42 46 50 52 59 60 66 68 69 72 76 78 82 86 89 90 98 p1:14 A 3 7 17 23 27 37 47 57 63 67 77 83 87 93 97 p1:15 A 7 13 17 27 33 37 43 47 53 57 67 73 77 87 97 p1:16 A 2 3 4 6 11 12 14 16 21 22 23 24 26 28 32 33 34 42 44 51 52 53 54 56 58 61 62 63 64 68 71 72 73 74 78 82 83 84 86 92 93 94 p1:17 A 1 2 4 8 12 13 14 18 22 24 31 32 34 36 38 41 42 43 44 46 48 52 54 62 64 66 72 74 76 81 82 84 88 91 92 94 96 98 p1:18 A 2 6 14 16 20 26 28 30 36 38 40 44 50 52 56 58 62 66 68 74 80 84 88 90 92 98 p1:19 A 0 4 8 10 12 18 22 24 32 34 42 46 48 54 60 64 70 72 76 78 82 86 94 96 p1:20 A 3 10 13 22 23 33 43 50 52 62 63 70 73 80 83 92 p1:21 A 0 2 12 20 30 32 40 42 53 60 72 82 90 93 p1:22 A 2 12 22 32 37 38 42 48 52 58 62 67 72 77 82 92 97 p1:23 A 2 7 8 12 17 18 22 27 28 32 42 47 52 57 62 68 72 78 82 87 88 92 98 p1:24 A 5 9 11 15 21 38 46 48 49 51 55 56 59 68 71 79 81 85 86 88 89 91 95 96 98 99 p1:25 A 1 6 8 16 18 19 25 26 28 29 31 35 36 39 41 45 58 61 65 66 69 75 76 78 p1:26 A 1 2 5 6 8 11 15 18 21 22 25 26 32 34 35 44 45 52 55 56 58 61 62 65 72 74 75 81 82 84 85 86 88 91 92 95 98 p1:27 A 4 5 12 14 15 16 24 25 28 31 35 36 38 41 42 45 46 48 51 54 55 64 65 66 68 71 75 76 78 85 94 95 96 p1:28 A 1 4 8 11 13 14 16 21 23 24 26 33 34 41 44 48 50 51 53 54 56 58 60 61 63 64 66 67 68 73 74 76 77 78 83 84 90 91 94 97 p1:29 A 0 3 4 6 7 10 14 17 18 20 24 27 28 30 31 34 36 37 38 40 43 44 46 47 54 57 64 70 71 74 80 81 84 86 87 88 93 94 96 98 p1:30 A 14 23 24 34 43 44 53 64 84 93 p1:31 A 3 4 13 33 54 63 73 74 83 94 p1:32 A 1 4 11 22 32 34 44 51 52 62 72 91 92 94 p1:33 A 2 12 14 21 24 31 41 42 54 61 64 71 74 81 82 84 p1:34 A 0 1 8 10 11 12 17 20 21 28 32 47 48 49 57 58 62 67 77 79 81 82 87 88 90 92 98 99 p1:35 A 2 7 9 18 19 22 27 29 30 31 37 38 39 40 41 42 50 51 52 59 60 61 68 69 70 71 72 78 80 89 91 97 p1:36 A 0 10 13 14 15 20 24 30 34 35 38 40 43 44 45 50 53 58 60 63 64 65 70 74 78 80 84 85 88 90 93 94 98 p1:37 A 0 3 4 5 8 10 18 20 23 25 28 30 33 40 48 50 54 55 60 68 70 73 75 80 83 90 95 p1:38 A 4 9 12 13 14 23 28 32 33 34 39 44 49 53 56 58 59 63 64 69 72 74 75 76 82 83 84 89 92 98 99 p1:39 A 2 3 5 6 8 15 16 18 19 22 24 25 26 29 35 36 38 42 43 45 46 48 52 54 55 62 65 66 68 73 78 79 85 86 88 93 94 95 96 p1:40 A 1 5 7 8 11 14 15 17 18 19 21 25 27 29 33 34 35 37 38 39 43 44 47 49 53 54 55 57 58 59 61 63 64 65 67 71 74 77 78 79 84 85 87 88 94 97 99 p1:41 A 3 4 7 9 13 17 23 24 27 28 31 37 41 45 47 48 51 57 67 68 69 73 75 77 81 83 87 89 91 93 95 97 98 p1:42 A 0 2 6 7 8 12 18 20 22 27 30 32 36 42 46 52 56 57 58 60 62 66 67 68 72 76 82 87 92 97 p1:43 A 2 10 12 16 17 22 26 28 32 37 38 40 42 47 48 50 52 62 70 72 77 78 80 82 86 88 90 92 96 98 p1:44 A 1 3 8 9 11 17 21 23 28 29 31 37 41 48 49 51 57 61 63 67 68 69 71 78 81 87 91 97 98 p1:45 A 1 7 11 13 18 19 21 27 31 33 38 39 41 43 47 51 53 58 59 61 71 73 77 79 81 83 88 89 91 93 99 p1:46 A 1 10 11 12 16 20 21 22 31 40 42 46 50 52 60 61 62 66 70 80 90 91 92 p1:47 A 0 2 6 26 30 32 36 41 51 56 71 72 76 81 82 86 96 p1:48 A 0 3 5 7 9 10 15 19 25 27 29 33 35 37 39 43 45 55 57 59 63 65 67 75 85 95 p1:49 A 5 13 15 17 20 23 25 30 35 40 45 47 49 50 53 55 60 65 69 70 73 75 77 79 80 83 85 87 89 90 93 95 97 99 p1:50 A 1 2 3 12 16 18 26 31 36 43 48 51 63 73 76 82 86 91 93 96 p1:51 A 6 8 11 13 21 22 23 28 32 33 38 41 42 46 52 53 56 58 61 62 66 68 71 72 78 81 83 88 92 98 p1:52 A 0 2 5 8 9 10 12 13 15 16 19 25 28 32 35 43 45 48 50 55 56 58 59 60 65 66 68 69 72 75 76 79 83 85 86 88 92 93 95 99 p1:53 A 3 5 6 15 18 20 22 23 25 26 29 30 33 35 36 38 39 40 42 45 46 49 52 53 55 62 63 65 70 73 75 78 80 82 85 89 90 95 96 98 p1:54 A 0 2 12 13 20 21 22 23 40 41 42 43 50 52 71 81 82 91 92 p1:55 A 1 3 10 11 30 31 32 33 51 53 60 61 62 63 70 72 73 80 83 90 93 p1:56 A 0 2 3 10 12 13 14 19 20 22 23 29 30 33 39 40 42 43 44 49 50 53 54 60 64 70 72 73 74 80 89 90 99 p1:57 A 0 4 9 10 20 24 30 32 34 40 50 52 59 60 62 63 69 70 79 80 82 83 84 90 92 93 94 p1:58 A 0 1 6 10 11 16 20 26 30 31 36 40 46 50 51 53 56 60 63 66 70 76 80 86 90 93 96 p1:59 A 0 3 6 10 13 16 20 21 23 26 30 33 36 40 41 43 46 50 56 60 61 66 70 71 73 76 80 81 83 86 90 91 96 p1:60 A 1 7 11 17 27 31 37 41 47 57 67 77 87 97 p1:61 A 7 17 21 27 37 47 51 57 61 67 71 77 81 87 91 97 p1:62 A 1 3 5 7 8 11 13 15 21 25 27 28 31 33 35 38 41 43 44 45 47 53 55 61 64 65 67 68 73 75 77 81 83 84 85 87 88 91 93 94 95 97 98 p1:63 A 4 5 14 15 17 18 23 24 25 34 35 37 45 48 51 54 55 57 58 63 65 71 74 75 78 85 95 p1:64 A 3 13 16 24 26 28 34 36 38 48 49 54 55 59 66 68 69 75 83 84 85 89 99 p1:65 A 4 5 6 8 9 14 15 18 19 23 25 29 33 35 39 43 44 45 46 53 56 58 63 64 65 73 74 76 78 79 86 88 93 94 95 96 98 p1:66 A 1 4 5 6 7 12 15 16 21 25 32 35 36 37 41 44 45 54 55 56 65 66 71 72 75 81 84 85 87 91 94 95 97 p1:67 A 2 5 11 14 15 17 22 24 25 26 27 31 34 35 42 45 46 47 51 52 55 57 61 62 64 65 67 74 75 76 77 82 85 86 92 95 96 p1:68 A 2 18 32 82 p1:69 A 8 12 22 28 38 42 48 52 58 62 68 72 78 88 92 98 p1:70 A 6 16 26 45 56 85 86 95 96 p1:71 A 5 15 25 35 36 46 55 65 66 75 76 p1:72 A 0 2 7 9 12 13 15 17 19 20 22 27 29 30 32 33 35 39 40 42 45 50 52 55 57 59 62 63 65 67 69 72 73 80 82 87 92 97 p1:73 A 2 3 5 10 12 22 23 25 32 37 42 43 47 49 52 53 60 62 70 72 75 77 79 82 83 85 89 90 92 93 95 99 p1:74 A 1 4 5 11 14 16 21 24 26 29 45 46 49 59 61 66 74 81 84 85 89 95 96 99 p1:75 A 6 9 15 19 25 31 34 35 36 39 41 44 51 54 55 56 64 65 69 71 75 76 79 86 91 94 p1:76 A 10 19 20 27 30 35 39 40 47 49 50 55 57 60 75 77 87 89 90 95 97 p1:77 A 0 5 7 9 15 17 25 29 37 45 59 65 67 69 70 79 80 85 99 p1:78 A 2 3 8 13 16 17 18 22 23 26 27 28 32 33 36 37 38 42 43 47 48 53 57 62 63 66 67 68 72 73 76 77 83 88 92 93 96 97 98 p1:79 A 3 6 7 12 13 23 33 43 46 52 53 56 58 63 73 78 82 83 86 87 93 p1:80 A 6 8 13 23 26 29 31 38 39 46 48 49 53 58 59 61 68 76 79 86 89 91 98 99 p1:81 A 1 3 9 11 16 18 19 21 28 33 36 41 43 51 56 63 66 69 71 73 78 81 83 88 93 96 p1:82 A 1 8 11 14 21 24 25 28 30 31 35 36 38 40 41 45 47 50 51 54 55 57 60 61 64 65 66 68 71 74 78 80 81 85 86 87 90 91 94 95 97 98 p1:83 A 0 1 4 5 6 7 10 11 15 16 17 18 20 21 26 27 31 34 37 41 44 46 48 51 56 58 61 67 70 71 75 76 77 81 84 88 91 96 p1:84 A 1 2 12 16 21 22 27 32 42 46 52 62 71 72 77 82 92 96 97 p1:85 A 2 6 7 11 12 17 22 26 31 32 36 37 41 42 47 51 52 56 57 61 62 66 67 72 76 81 82 86 87 91 92 p1:86 A 2 9 22 31 32 39 41 42 49 59 60 62 69 71 80 82 89 92 99 p1:87 A 0 1 10 11 12 19 20 21 29 30 40 50 51 52 61 70 72 79 81 90 91 p1:88 A 1 2 3 4 5 13 14 15 16 21 22 23 25 31 33 34 35 36 37 41 43 45 46 47 51 52 53 55 57 61 63 65 67 71 73 74 75 77 81 83 84 85 86 91 92 93 95 96 p1:89 A 3 5 6 7 11 12 13 15 17 23 24 25 26 27 32 33 35 42 43 44 45 53 54 55 56 62 63 64 65 66 72 73 75 76 82 83 85 87 93 94 95 97 p1:90 A 4 5 8 9 10 11 14 15 20 24 31 34 39 40 41 46 48 49 51 55 58 60 64 65 66 69 70 71 76 79 80 81 86 89 94 95 98 p1:91 A 0 1 6 16 18 19 21 25 26 28 29 30 35 36 38 44 45 50 54 56 59 61 68 74 75 78 84 85 88 90 91 96 99 p1:92 A 1 6 11 16 21 31 36 41 51 56 61 71 81 86 91 p1:93 A 1 11 21 26 31 41 46 51 61 66 71 76 81 91 96 p1:94 A 0 3 10 13 15 18 20 21 23 25 29 30 31 38 40 45 49 50 51 53 59 60 61 63 68 69 70 71 75 78 79 80 81 88 90 91 93 95 98 99 p1:95 A 0 1 5 8 9 10 11 19 20 28 30 33 35 39 40 41 43 48 50 55 58 60 65 70 73 80 83 85 89 90 p1:96 A 3 4 5 13 15 23 25 26 33 34 43 46 53 54 55 63 64 65 66 73 75 83 84 86 93 95 p1:97 A 3 6 13 14 16 23 24 33 35 36 43 44 45 53 56 63 73 74 76 83 85 93 94 96 p1:98 A 7 13 16 19 26 27 35 37 45 65 66 67 75 85 86 87 93 95 p1:99 A 3 5 6 9 15 17 23 25 29 33 36 39 43 46 47 49 53 55 56 57 59 63 69 73 76 77 79 83 89 96 97 99 p1:100 A 1 2 12 13 15 21 23 25 35 43 48 52 53 55 62 63 65 68 71 72 78 81 85 93 95 p1:101 A 3 5 8 11 18 22 28 31 32 33 38 41 42 45 51 58 61 73 75 82 83 88 91 92 98 p1:102 A 2 3 5 6 11 12 13 15 21 22 23 25 26 33 36 41 42 43 53 55 62 63 65 66 71 73 75 83 85 91 92 93 95 96 p1:103 A 1 3 13 16 23 31 32 33 35 43 45 46 51 52 53 56 61 63 72 73 76 81 82 83 86 93 p1:104 A 2 12 22 26 32 42 52 56 62 72 82 92 96 p1:105 A 2 6 12 16 22 32 36 42 46 52 62 66 72 76 82 86 92 p1:106 A 29 49 59 79 89 99 p1:107 A 9 19 39 69 p1:108 A 1 2 11 12 13 21 22 28 31 32 38 41 42 43 48 51 52 53 61 62 68 71 72 73 74 81 82 83 91 92 93 p1:109 A 1 2 3 4 8 11 12 14 18 21 22 23 24 31 32 33 34 41 42 44 51 52 54 58 61 62 63 64 71 72 78 81 82 84 88 91 92 94 98 p1:110 A 3 7 9 13 15 23 25 27 29 33 39 43 44 53 54 55 59 63 65 73 79 83 93 94 95 p1:111 A 3 4 5 13 14 17 19 23 24 33 34 35 37 43 45 47 49 53 57 63 64 67 69 73 74 75 77 83 84 85 87 89 93 97 99 p1:112 A 1 2 5 8 11 12 21 22 41 42 52 55 61 68 71 78 82 88 91 92 95 98 p1:113 A 15 18 25 28 31 32 35 38 45 48 51 58 62 65 72 75 81 85 p1:114 A 1 2 3 9 13 15 19 25 29 31 32 39 42 49 54 55 59 61 62 69 71 72 73 79 82 84 89 91 95 99 p1:115 A 4 5 9 11 12 14 19 21 22 23 24 29 33 34 35 39 41 43 44 45 49 51 52 53 59 63 64 65 69 74 75 79 81 83 85 89 92 93 94 99 p1:116 A 8 14 15 18 24 28 35 38 44 48 54 55 58 65 68 74 78 88 98 p1:117 A 4 5 8 18 25 28 34 38 45 48 58 64 68 75 78 84 85 88 94 95 98 p1:118 A 23 26 33 37 43 46 47 53 56 67 70 76 77 83 87 90 93 96 97 p1:119 A 0 3 6 7 10 13 16 17 20 27 30 36 40 50 57 60 63 66 73 80 86 p1:120 A 2 6 7 11 12 16 17 19 26 27 28 31 36 37 42 46 47 51 52 56 57 66 67 68 69 71 72 76 77 78 79 81 82 86 87 88 89 91 92 96 97 p1:121 A 1 6 7 8 9 16 17 18 21 22 26 27 29 32 36 37 38 39 41 46 47 48 49 56 57 58 59 61 62 66 67 76 77 86 87 96 97 98 99 p1:122 A 1 3 13 31 33 41 51 53 61 63 81 91 93 p1:123 A 11 21 23 43 71 73 83 p1:124 A 0 9 17 19 20 23 24 34 35 37 43 53 55 60 64 65 77 79 80 83 90 93 97 p1:125 A 3 4 5 7 10 13 14 15 25 27 29 30 33 39 40 44 45 47 49 50 54 57 59 63 67 69 70 73 74 75 84 85 87 89 94 95 99 p1:126 A 1 6 8 14 16 17 18 24 26 28 31 34 36 37 38 41 46 51 54 56 57 58 61 66 74 76 77 78 81 86 87 91 94 96 98 p1:127 A 4 6 7 11 16 21 26 27 36 44 46 47 48 56 64 66 67 68 71 76 84 86 88 96 97 p1:128 A  p1:129 A  p1:130 A 1 2 4 7 12 14 17 21 22 24 27 31 32 37 41 42 44 47 52 54 57 61 62 64 67 72 77 82 87 92 94 97 p1:131 A 2 7 11 12 17 22 27 32 34 37 42 47 51 52 57 62 67 71 72 74 77 81 82 84 87 91 92 97 p1:132 A 0 3 9 10 13 19 23 29 30 33 40 43 49 50 53 63 69 73 79 80 83 93 99 p1:133 A 3 13 20 23 33 39 43 53 59 60 63 70 73 83 89 90 93 p1:134 A 12 24 34 35 44 54 55 62 64 72 74 82 84 85 92 p1:135 A 2 4 5 14 15 22 25 32 42 45 52 65 75 94 95 p1:136 A 17 27 32 37 52 57 58 62 67 68 72 82 83 93 97 p1:137 A 2 3 7 8 12 13 18 22 23 28 33 38 42 43 47 48 53 63 73 77 78 87 88 92 98 p1:138 A 0 1 3 5 11 13 15 18 20 21 22 23 25 28 30 31 32 33 35 38 41 43 45 50 51 52 53 55 61 62 63 65 73 75 80 82 83 85 90 91 93 95 p1:139 A 2 3 5 8 10 12 13 15 23 25 33 35 40 42 43 45 48 53 55 58 60 63 65 68 70 71 72 73 75 78 81 83 85 88 92 93 95 98 p1:140 A 2 4 8 9 14 16 18 19 24 28 32 35 36 39 44 46 48 49 52 54 58 62 66 72 79 82 86 88 94 95 99 p1:141 A 5 6 12 15 22 25 26 29 34 38 42 45 55 56 59 64 65 68 69 74 75 76 78 84 85 89 92 96 98 p1:142 A 0 1 2 6 7 10 11 16 17 20 21 26 27 30 31 34 36 37 40 41 46 47 50 51 56 57 60 61 62 64 66 67 70 71 74 76 77 80 81 82 86 87 90 91 96 97 p1:143 A 0 1 4 6 7 10 11 12 14 16 17 20 21 22 24 26 27 30 31 32 36 37 40 41 42 44 46 47 50 51 52 54 56 57 60 61 66 67 70 71 72 76 77 80 81 84 86 87 90 91 92 94 96 97 p1:144 A 7 14 15 16 17 24 25 34 45 46 47 54 55 56 57 64 65 66 74 76 77 84 85 86 94 95 97 99 p1:145 A 4 5 6 9 19 26 27 29 35 36 37 39 44 49 59 67 69 75 79 87 89 96 p1:146 A 13 23 33 43 63 83 93 p1:147 A 3 53 73 p1:148 A 3 10 13 20 21 23 24 28 33 43 48 50 53 54 58 60 63 67 68 71 73 78 80 81 84 87 93 94 97 98 p1:149 A 0 1 4 7 8 11 14 17 18 27 30 31 34 37 38 40 41 44 47 51 57 61 64 70 74 77 83 88 90 91 p1:150 A 1 2 5 9 13 15 19 20 21 23 30 37 42 43 45 50 52 53 57 60 63 67 69 73 81 85 87 91 92 93 97 99 p1:151 A 0 3 7 10 11 12 17 22 25 27 29 31 32 33 35 39 40 41 47 49 51 55 59 61 62 65 70 71 72 75 77 79 80 82 83 89 90 95 p1:152 A 6 9 14 16 19 27 29 36 37 39 46 47 49 54 56 57 59 64 69 76 79 86 87 89 99 p1:153 A 4 7 9 17 19 24 26 29 34 39 44 49 59 66 67 69 74 77 79 84 89 94 96 97 99 p1:154 A 4 8 11 13 15 18 21 24 34 38 43 44 48 53 55 63 64 65 68 71 75 78 81 83 84 88 91 93 94 95 p1:155 A 1 3 5 14 23 25 28 31 33 35 41 45 51 54 58 61 73 74 85 98 p1:156 A 0 6 13 20 30 33 36 38 41 43 46 48 56 58 71 73 76 78 83 86 88 93 96 p1:157 A 1 3 8 10 11 16 18 21 23 26 28 31 40 50 51 53 60 61 63 66 68 70 80 81 90 91 98 p1:158 A 2 3 5 6 8 12 15 19 23 25 26 29 32 33 35 38 39 42 43 45 46 49 52 55 56 59 62 65 68 73 75 76 78 79 82 83 85 86 89 92 95 98 99 p1:159 A 5 9 13 15 16 18 22 25 28 35 36 45 48 53 55 58 63 65 66 69 72 75 85 88 93 95 96 p1:160 A 1 11 19 21 29 31 39 41 49 51 59 61 71 81 89 91 99 p1:161 A 1 9 11 21 31 41 51 61 69 71 79 81 91 p1:162 A 0 7 8 10 17 18 22 27 28 30 32 37 38 47 48 52 57 58 60 67 68 77 78 82 87 88 90 97 98 p1:163 A 2 7 8 12 17 18 20 27 28 37 38 40 42 47 48 50 57 58 62 67 68 70 72 77 78 80 87 88 92 97 98 p1:164 A 0 1 10 11 15 21 23 26 30 31 33 40 41 43 50 51 61 63 70 71 73 81 83 86 90 91 93 96 p1:165 A 1 3 5 6 11 13 16 20 21 25 31 35 36 41 45 46 51 53 55 56 60 61 65 66 71 75 76 80 81 85 91 95 p1:166 A 3 18 19 28 38 49 53 63 69 73 78 83 93 98 99 p1:167 A 8 9 13 23 29 33 39 43 48 58 59 68 79 88 89 p1:168 A 4 5 6 7 8 12 14 15 18 24 26 27 28 29 32 34 38 44 45 46 48 49 52 54 55 57 58 62 64 67 68 69 72 74 76 77 78 82 84 85 87 88 89 92 94 96 97 98 p1:169 A 2 4 8 9 14 16 17 18 19 22 24 25 28 34 35 36 37 38 39 42 44 47 48 54 56 58 59 64 65 66 68 74 75 78 79 84 86 88 94 95 98 99 p1:170 A 5 15 16 25 26 36 39 45 59 65 66 69 75 79 85 p1:171 A 6 9 19 29 35 46 49 55 56 76 86 89 95 96 99 p1:172 A 1 11 14 15 21 30 31 34 35 40 41 50 51 54 60 61 64 80 85 90 94 p1:173 A 0 4 5 10 20 24 25 44 45 55 65 70 71 74 75 81 84 91 95 p1:174 A 6 9 15 16 25 26 29 36 46 47 55 56 57 59 65 66 67 75 76 85 86 95 96 97 99 p1:175 A 5 6 7 16 17 19 26 27 35 36 37 39 45 46 49 56 66 69 76 77 79 86 87 89 96 p1:176 A 70 80 90 p1:177 A 0 10 20 30 40 50 60 p1:178 A 4 12 14 22 24 25 34 42 44 45 46 54 56 64 65 66 72 74 82 84 85 86 92 95 96 p1:179 A 2 5 6 15 16 26 32 35 36 52 55 62 75 76 94 p1:180 A 0 1 4 8 10 11 14 20 21 25 30 31 35 41 44 48 51 54 55 60 61 64 70 71 74 81 84 88 91 98 p1:181 A 1 5 11 15 18 21 24 28 31 34 38 40 41 45 50 51 58 61 65 68 71 75 78 80 81 85 90 91 94 95 p1:182 A 0 2 6 12 22 32 36 40 46 50 62 76 82 90 92 p1:183 A 10 16 20 26 30 42 52 56 60 66 70 72 80 86 96 p1:184 A 0 1 3 5 10 11 13 15 19 20 21 22 23 25 28 29 33 40 48 51 52 53 55 60 62 63 71 72 73 79 83 88 90 92 93 95 98 p1:185 A 2 8 9 12 18 30 31 32 35 38 39 41 42 43 45 49 50 58 59 61 65 68 69 70 75 78 80 81 82 85 89 91 99 p1:186 A 4 5 14 15 17 24 25 29 35 37 45 47 49 54 59 65 69 74 75 77 85 87 89 99 p1:187 A 7 9 19 27 34 39 44 55 57 64 67 79 84 94 95 97 p1:188 A 1 2 7 10 11 12 17 20 21 22 31 32 40 42 45 52 57 61 62 71 72 77 80 82 85 90 91 92 97 p1:189 A 0 2 5 12 15 22 25 27 30 32 35 37 41 42 47 50 51 52 55 60 62 65 67 70 72 75 81 82 87 92 95 p1:190 A 0 5 6 7 10 15 16 17 20 25 26 27 29 30 35 36 37 40 45 46 47 50 55 56 57 59 60 65 66 67 69 70 75 76 77 79 80 85 86 87 89 90 95 96 97 99 p1:191 A 0 5 6 7 9 10 15 16 17 19 20 25 26 27 30 35 36 37 39 40 45 46 47 49 50 55 56 57 60 65 66 67 70 75 76 77 80 85 86 87 90 95 96 97 p1:192 A 1 11 21 28 31 41 51 61 68 71 81 88 91 98 p1:193 A 1 8 11 18 21 31 38 41 48 51 58 61 71 78 81 91 p1:194 A 5 8 11 15 18 21 22 25 31 32 35 38 39 45 55 59 61 62 65 68 69 75 78 82 85 92 95 98 p1:195 A 1 2 5 9 12 15 19 25 28 29 35 41 42 45 48 49 51 52 55 58 65 71 72 75 79 81 85 88 89 91 95 99 p1:196 A 0 1 4 5 8 20 24 25 26 31 34 36 41 48 51 56 58 60 61 64 65 66 70 76 78 80 81 84 86 88 95 96 98 p1:197 A 6 10 11 14 15 16 18 21 28 30 35 38 40 44 45 46 50 54 55 68 71 74 75 85 90 91 94 p1:198 A 0 1 3 10 11 12 13 19 21 23 29 30 33 34 39 40 41 43 44 50 51 53 60 61 63 64 69 71 73 74 79 81 83 84 89 91 93 p1:199 A 2 3 4 9 13 14 20 22 23 24 31 32 33 42 43 49 52 53 54 59 62 63 70 72 73 80 82 83 90 92 93 94 99