Create a file to initialise a population in SLiM

Make a SLiM file in the style of outputFull() which can be read by SLiM. This is useful to start a simulation with a particular population state that you define in R

slim_make_pop_input(
  snps,
  file_name = tempfile(),
  sim_gen = 10000,
  ind_pops = NULL,
  ind_sex = NULL,
  mut_pos = NULL,
  mut_type = NULL,
  mut_sel = NULL,
  mut_dom = NULL,
  mut_pop = NULL,
  mut_gen = NULL,
  mut_nuc = NULL
)

Arguments

snps
R object containing SNPs. Can be a matrix with 0, 1, or 2 as its elements, or a genlight object. Rows correspond to individuals, columns to loci.

file_name
Path to file where the population input file should be saved.

sim_gen
What generation to write in the file. This doesn't really do anything, but is just a requirement for the format.

ind_pops
An optional character vector with length equal to nrow(snps) with Subpopulation indicators for each individual. Should be "p1", "p2", etc. as this is how subpopulations are named in SLiM.

ind_sex
An optional character vector with length equal to nrow(snps) with the sex for each individual (can be "H" for hermaphrodite, "F" for female, or "M" for male.)

mut_pos
An optional integer vector with length equal to ncol(snps) specifying the positions in the genome of each loci.

mut_type
An optional character vector with length equal to nrow(snps) with the mutation type of each loci. Should be "m1", "m2", etc. as this is how mutation types are named in SLiM.

mut_sel
An optional numeric vector with length equal to nrow(snps) with the selection coefficient for the mutation type of each loci.

mut_dom
An optional numeric vector with length equal to nrow(snps) with the dominance coefficient for the mutation type of each loci.

mut_pop
An optional character vector with length equal to nrow(snps) with the subpopulation of origin for the mutation at each loci. Should be "p1", "p2", etc. as this is how subpopulations are named in SLiM.

mut_gen
An optional integer vector with length equal to ncol(snps) specifying the generation of origin for the mutation at each loci.

mut_nuc
An optional character vector with length equal to nrow(snps) with the nucleotide of the mutation. Should be "A", "C", "G", or "T".

Value